+++ bookstore_server/src/main/java/com/huang/store/mapper/AddressMapper.java
@@ -8,7 +8,7 @@ import java.util.List;
 @Repository
 public interface AddressMapper {
     int addAddress(Address address);
-    int deleteAddress(int id);
+    int deleteAddress(Long id);
     int modifyAddress(Address address);
     List<Address> addressList(String account);
     int setAddressOff(int id);//设置地址为删除状态
