+++ bookstore_server/src/main/java/com/huang/store/mapper/BookMapper.java
@@ -11,21 +11,21 @@ public interface BookMapper {
     //对于book表的操作
     int addBook(Book book);
     int modifyBook(Book book);
-    int modifyBookPut(int id,boolean put);//修改图书是否为上下架
-    int modifyBookRec(int id,boolean recommend);//修改图书是否为推荐图书
-    int modifyBookNewPro(int id,boolean newProduct);//修改图书是否为新品
-    int modifyBookStock(int id,int stockNum);//减库存stockNum
-    int deleteBook(int id);//删除图书
+    int modifyBookPut(Long id,boolean put);//修改图书是否为上下架
+    int modifyBookRec(Long id,boolean recommend);//修改图书是否为推荐图书
+    int modifyBookNewPro(Long id,boolean newProduct);//修改图书是否为新品
+    int modifyBookStock(Long id,int stockNum);//减库存stockNum
+    int deleteBook(Long id);//删除图书
     List<Book> getBooks();//得到所有图书
     List<Book> getBooksByPage(int page,int pageSize);//按页得到图书集合
     List<Book> getNewPutBookList(int page,int pageSize);//按页得到新上架的图书集合
 
-    List<OrderBookDto> getBatchBookList(int[] ids);//根据ids数组，得到对应的图书集合
-    List<OrderBookDto> getOneBookList(int[] ids);//根据ids数组，得到对应的图书集合
+    List<OrderBookDto> getBatchBookList(Long[] ids);//根据ids数组，得到对应的图书集合
+    List<OrderBookDto> getOneBookList(Long[] ids);//根据ids数组，得到对应的图书集合
 
-    int getBookId(String isbn);
-    String getBookIsbn(int id);
-    Book getBook(int id);//获取图书的所有
+    Long getBookId(String isbn);
+    String getBookIsbn(Long id);
+    Book getBook(Long id);//获取图书的所有
     Book getBookDetail(String isbn);//获得图书用于后台页表展示的信息(除去图书详情页和部分图书相册)
     int getPublishBookNum(String publishName);//得到某一出版社的图书的数量
     int getBookCount();//得到图书的数量
@@ -43,29 +43,29 @@ public interface BookMapper {
 
     //对于推荐图书的操作
     int addToRecommend(Recommend recommend);
-    int deleteFromRecommend(int bookId);
-    int modifyRecommendRank(int bookId,int rank);
+    int deleteFromRecommend(Long bookId);
+    int modifyRecommendRank(Long bookId,int rank);
     int modifyRecommend(Recommend recommend);
-    int hasExistInRec(int bookId);
+    int hasExistInRec(Long bookId);
     List<Book> getRecommendsByPage(int page,int pageSize);
 
     //对于新品推荐的操作
     int addToNewProduct(Recommend newProduct);
-    int deleteFromNewProduct(int bookId);
-    int modifyNewProductRank(int bookId,int rank);
+    int deleteFromNewProduct(Long bookId);
+    int modifyNewProductRank(Long bookId,int rank);
     int modifyNewProduct(Recommend newProduct);
-    int hasExistInNew(int bookId);
+    int hasExistInNew(Long bookId);
     List<Book> getNewProductsByPage(int page,int pageSize);
 
 
     //添加图书到分类
-    int addBookToSort(int bookSortId ,int bookId);
-    int delBookFromSort(int booSortId,int bookId);
-    int modifyBookSort(int bookSortId,int bookId);
-    BookSort getBookSort(int bookId);//得到某本书的分类名
-    List<Book> getBooksByFirst(List<Integer> list, int page, int pageSize);//得到级别一的所有分类书籍
-    List<Book> getBookBySecond(int bookSortId,int page,int pageSize);//得到级别二的所有分类图书
-    int getFirstBookCount(List<Integer> list);//得到一级分类图书的数量
-    int getSecondBookCount(int bookSortId);//得到二级分类图书的数量
+    int addBookToSort(Long bookSortId ,Long bookId);
+    int delBookFromSort(Long booSortId,Long bookId);
+    int modifyBookSort(Long bookSortId,Long bookId);
+    BookSort getBookSort(Long bookId);//得到某本书的分类名
+    List<Book> getBooksByFirst(List<Long> list, int page, int pageSize);//得到级别一的所有分类书籍
+    List<Book> getBookBySecond(Long bookSortId,int page,int pageSize);//得到级别二的所有分类图书
+    int getFirstBookCount(List<Long> list);//得到一级分类图书的数量
+    int getSecondBookCount(Long bookSortId);//得到二级分类图书的数量
 }
