+++ bookstore_server/src/main/java/com/huang/store/service/imp/AddressService.java
@@ -8,7 +8,7 @@ import java.util.List;
 @Service
 public interface AddressService {
     int addAddress(Address address);
-    int deleteAddress(int id);
+    int deleteAddress(Long id);
     int modifyAddress(Address address);
     List<Address> addressList(String account);
     int setAddressOff(int id);
