+++ bookstore_server/src/main/java/com/huang/store/service/imp/OrderService.java
@@ -21,17 +21,17 @@ public interface OrderService {
 
     boolean addOrder(OrderInitDto orderInitDto);//添加订单 其中涉及添加总的订单表、订单明细表、费用信息表、还有减去库存的操作等
 
-    int delOrder(int id);//删除订单  这里涉及了多表删除  并且只能删除已经完成了的订单
+    int delOrder(Long id);//删除订单  这里涉及了多表删除  并且只能删除已经完成了的订单
 
-    int userDelOrder(int id);//用户删除订单,这里的删除不是真正的删除，而是设置删除标志位为true
+    int userDelOrder(Long id);//用户删除订单,这里的删除不是真正的删除，而是设置删除标志位为true
 
     int batchDelOrder(List<Integer> item);//批量删除 这里涉及多表联合删除
 
-    int modifyOrderStatus(int id, String orderStatus);//修改订单的状态
+    int modifyOrderStatus(Long id, String orderStatus);//修改订单的状态
 
-    int deliverBook(int id,int logisticsCompany,String logisticsNum);//发货
+    int deliverBook(Long id,int logisticsCompany,String logisticsNum);//发货
 
-    OrderDto findOrderDto(int id);//得到某个订单的所有信息明细
+    OrderDto findOrderDto(Long id);//得到某个订单的所有信息明细
 
     List<OrderDetailDto> findOrderDetailDtoList(String orderId);//批量得到指定订单所有图书的明细信息
